+++ src/core/io/avd/avd_parser.hpp
@@ -1,0 +1,36 @@
+#pragma once
+
+#include <memory>
+#include <functional>
+
+#include "io/base.hpp"
+
+namespace glaxnimate::io::avd {
+
+class AvdParser
+{
+private:
+
+
+public:
+    /**
+     * \throws SvgParseError on error
+     */
+    AvdParser(
+        QIODevice* device,
+        model::Document* document,
+        const std::function<void(const QString&)>& on_warning = {},
+        ImportExport* io = nullptr,
+        QSize forced_size = {}
+    );
+
+    ~AvdParser();
+
+    void parse_to_document();
+
+    class Private;
+private:
+    std::unique_ptr<Private> d;
+};
+
+} // namespace glaxnimate::io::avd